Changed prodlim to use long doubles - no real difference

+++ src/prodlim.h
@@ -9,16 +9,17 @@ class prodlim{
     std::vector<int> trun;
     std::vector<double> lambda_0;
     std::vector<double> lambda_1;
-    std::vector<double> risk_0;
+    std::vector<long double> risk_0;
     std::vector<int> left_full;
     std::vector<int> right_full;
     std::vector<int> trun_full;
 
     std::vector<double> c;
     std::vector<double> cum_lambda;
-    std::vector<double> surv;
-    std::vector<double> p_obs;
-    std::vector<double> n_trans, cum_n_trans, h, w_sum;
+    std::vector<long double> surv;
+    std::vector<long double> p_obs;
+    std::vector<long double> n_trans, cum_n_trans, h, w_sum;
+    
 
     class invert_data{
       public:
@@ -57,9 +58,9 @@ class prodlim{
       right = Rcpp::as< std::vector<int> >(r);
       trun = Rcpp::as<std::vector<int> >(t);
       lambda_0 = Rcpp::as< std::vector<double> >(lambda);
-      h = Rcpp::as< std::vector<double> >(lambda);
+      h = Rcpp::as< std::vector<long double> >(lambda);
       lambda_1 = lambda_0;
-      risk_0 = Rcpp::as< std::vector<double> >(R0);
+      risk_0 = Rcpp::as< std::vector<long double> >(R0);
       left_full = Rcpp::as< std::vector<int> >(l_full);
       right_full = Rcpp::as< std::vector<int> >(r_full);
       trun_full = Rcpp::as<std::vector<int> >(t_full);
